+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 2.0.0rc2 (unreleased)
 ---------------------
 
+- #1655 Rename service's "Result Options" and "Additional Values"
+- #1655 Move service's "Additional values" to "Result Options" tab
 - #1650 Fix Error when invalidating a sample with contained retests
 - #1646 Allow multi-select in results entry
 - #1645 Allow translation of path bar items